requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-authorization
